Extract VAE orchestration into manager
